requirements.txt
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
langchain==0.2.14
langchain-openai==0.1.25
langchain-chroma==0.1.2
chromadb==0.5.5
tiktoken==0.7.0
streamlit==1.38.0
PyYAML==6.0.2
